add async/await

+++ node-todo-api/server/server.js
@@ -14,69 +14,57 @@ const port = process.env.PORT
 
 app.use(bodyParser.json())
 
-app.post('/todos', authenticate, (req, res) => {
+app.post('/todos', authenticate, async (req, res) => {
   const todo = new Todo({
     text: req.body.text,
     _creator: req.user._id
   })
 
-  todo.save()
-    .then(todo => {
-      res.send(todo)
-    })
-    .catch(err => {
-      res.status(400).send(err)
-    })
-
-})
-
-app.get('/todos', authenticate, (req, res) => {
-  Todo.find({
-    _creator: req.user._id
-  })
-  .then(todos => {
-    res.send({todos})
-  })
-  .catch(err => {
+  try {
+    await todo.save()
+    res.status(200).send(todo)
+  } catch(err) {
     res.status(400).send(err)
-  })
+  }
 })
 
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos', authenticate, async (req, res) => {
+
+  try {
+    const todos = await Todo.find({_creator: req.user._id})
+    res.status(200).send({todos})
+  } catch(err) {
+    res.status(400).send(err)
+  }
+})
+
+app.get('/todos/:id', authenticate, async (req, res) => {
   const {id} = req.params
   if (!ObjectId.isValid(id)) return res.status(400).send('Invalid id')
 
-  Todo.findOne({
-    _id: id,
-    _creator: req.user._id
-  })
-  .then(todo => {
-    if(!todo) return res.status(400).send('No match found')
+  try {
+    const todo = await Todo.findOne({_id: id,_creator: req.user._id})
+    if(!todo) throw new Error('No match found')
     res.send(todo)
-  })
-  .catch(err => {
+  } catch(err) {
     res.status(400).send(err)
-  })
+  }
 })
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
   const {id} = req.params
   if (!ObjectId.isValid(id)) return res.status(400).send('Invalid id')
 
-  Todo.findOneAndDelete({
-    _id: id,
-    _creator: req.user._id
-  })
-  .then(todo => {
-    if(!todo) return res.status(400).send('No match found')
+  try {
+    const todo = await Todo.findOneAndDelete({_id: id,_creator: req.user._id})
+    if (!todo) throw new Error('No match found')
     res.send({todo})
-  })
-  .catch(err => {
+  } catch(err) {
     res.status(400).send(err)
-  })
+  }
 })
 
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
   const {id} = req.params
   const {body} = req
 
@@ -89,61 +77,55 @@ app.patch('/todos/:id', authenticate, (req, res) => {
     body.completedAt = null
   }
 
-  Todo.findOneAndUpdate({
-    _id: id,
-    _creator: req.user._id
-  }, {$set: body}, {new: true})
-  .then(todo => {
-    if (!todo) return res.status(400).send()
+  try {
+    const todo = await Todo.findOneAndUpdate({
+      _id: id,
+      _creator: req.user._id
+    }, {$set: body}, {new: true})
+
+    if (!todo) throw new Error('Item not found')
     res.send({todo})
-  })
-  .catch(err => {
+  } catch(err) {
     res.status(400).send()
-  })
+  }
 })
 
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
   const {email, password} = req.body
   const user = new User({email, password})
 
-  user.save()
-    .then(() => {
-      return user.generateAuthToken()
-    })
-    .then(token => {
-      res.header('x-auth', token)
-      .send(user)
-    })
-    .catch(err => {
-      res.status(400).send(err)
-    })
+  try {
+    await user.save()
+    const token = await user.generateAuthToken()
+    res.header('x-auth', token).send(user)
+  } catch(err) {
+    res.status(400).send(err)
+  }
 })
 
-app.post('/users/login', (req, res) => {
+app.post('/users/login', async (req, res) => {
   const {email, password} = req.body
 
-  User.findByCredentials(email, password)
-    .then(user => {
-      return user.generateAuthToken().then(token => {
-        res.header('x-auth', token)
-        .send(user)
-      })
-    })
-    .catch(err => {
-      res.status(401).send(err)
-    })
+  try {
+    const user = await User.findByCredentials(email, password)
+    const token = await user.generateAuthToken()
+    res.header('x-auth', token).send(user)
+  } catch(err) {
+    res.status(401).send(err)
+  }
 })
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user)
 })
 
-app.delete('/users/me/token', authenticate, (req,res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req,res) => {
+  try {
+    await req.user.removeToken(req.token)
     res.status(200).send()
-  }, () => {
+  } catch (err) {
     res.status(400).send()
-  })
+  }
 })
 
 app.listen(port, () => {
